+++ services/JourneyService/Journey.Application/DTOs/Response/MonthlyRouteDistanceResponse.cs
@@ -4,5 +4,5 @@ public sealed class MonthlyRouteDistanceResponse
 {
     public Guid UserId { get; set; }
     public string? FullName { get; set; }
-    public double TotalDistanceKm { get; set; }
+    public decimal TotalDistanceKm { get; set; }
 }
